requirements.txt
asgiref==3.9.1
astroid==2.15.8
black==24.3.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
cloudinary==1.44.1
dill==0.4.0
dj-database-url==3.0.1
Django==5.0.14
django-cloudinary-storage==0.3.0
django-cors-headers==4.3.1
django-extensions==4.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
gunicorn==21.2.0
idna==3.10
isort==5.13.2
lazy-object-proxy==1.12.0
mccabe==0.7.0
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
pillow==11.3.0
platformdirs==4.4.0
psycopg==3.2.9
psycopg-binary==3.2.9
PyJWT==2.10.1
pylint==2.17.7
pylint-django==2.5.3
pylint-plugin-utils==0.9.0
python-decouple==3.8
python-dotenv==1.0.1
pytz==2025.2
requests==2.32.5
six==1.17.0
sqlparse==0.5.3
tomlkit==0.13.3
typing_extensions==4.15.0
urllib3==2.5.0
whitenoise==6.9.0
wrapt==1.17.3
